created a method in Hardware called webcamReg
